Refresh screen after running `!npm repo`
Since `npm repo` prints output, it causes Vim to show a corrupted buffer
after it finishes. There's an issue with a bit of detail here:
https://github.com/vim/vim/issues/1253

The solution is to run `redraw!` so that Vim repaints the screen and the
buffer looks just the same as it did before running `!npm repo`.